+++ src/operations.h
@@ -1,3 +1,4 @@
+#include <cstdint>
 #include <string>
 #include <vector>
 
@@ -8,8 +9,8 @@ public:
   static void Log(std::string msg, logLevel level);
 };
 
-std::string GetVectorAdditionOp(int size);
-std::string GetDotProductOp(int size);
-std::string GetTransposeOp(std::vector<int> shape);
-std::string GetMatrixMultiplicationOp(std::vector<int> matrix1Shape,
-                                      std::vector<int> matrix2Shape);
+std::string GetVectorAdditionOp(int64_t size);
+std::string GetDotProductOp(int64_t size);
+std::string GetTransposeOp(std::vector<int64_t> shape);
+std::string GetMatrixMultiplicationOp(std::vector<int64_t> matrix1Shape,
+                                      std::vector<int64_t> matrix2Shape);
